Move implementation out of entry file
